open_weather_api/requirements.txt
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.1
pylint==2.12.2
python-dotenv==0.20.0
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
wrapt==1.13.3
